Refactors to use minimal definition
In Rust we would only require a definition for the Next method and have
the rest of the methods filled in if necessary. Unfortunately we cannot
define methods on interfaces in Go so instead we embed and interface
into a struct. This might make the Iterator interface difficult to use
but I am unsure yet what the results will be.